refactored profile/code for readability. test for get_profile's failing
